Перенес логику получения изображения в сервисы
