+++ lib/features/home/service/service.dart
@@ -2,4 +2,6 @@
 // generated by: https://marketplace.visualstudio.com/items?itemName=elmsec.dart-flutter-exports:
 
 
+export 'get_difficulty_image.service.dart';
+export 'get_priority_image.service.dart';
 export 'task_sort.service.dart';
